+++ src/services/index.ts
@@ -1,4 +1,5 @@
 export * from "./auth-events";
 export * from "./auth.service";
+export * from "./current-user";
 export * from "./generate-jwt-secret";
 export * from "./jwt";